refactor: move: searchTag 패키지 이동
model -> model/studyaddons 로 이동

[skip ci]
