+++ server/edusync/src/main/java/com/codestates/edusync/model/study/studygroup/service/StudygroupService.java
@@ -6,7 +6,7 @@ import com.codestates.edusync.model.common.utils.VerifyVerifyStudygroupUtils;
 import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
 import com.codestates.edusync.model.study.studygroup.repository.StudygroupRepository;
 import com.codestates.edusync.model.study.studygroup.utils.StudygroupManager;
-import com.codestates.edusync.model.searchtag.service.SearchTagService;
+import com.codestates.edusync.model.studyaddons.searchtag.service.SearchTagService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
